Some x86 improvements and GDB stub support
git-svn-id: http://pearcolator.svn.sourceforge.net/svnroot/pearcolator@77 f9fe5334-281c-0410-a981-a1a969458ad4

+++ src/org/binarytranslator/generic/execution/GdbController.java
@@ -506,7 +506,7 @@ public class GdbController extends
    */
   private void handle_getMemValue(int dataEnd) throws IOException {
     String data = bufferToString(2, dataEnd);
-    int address = Integer.parseInt(data.substring(0, data.indexOf(',')), 16);
+    int address = (int)Long.parseLong(data.substring(0, data.indexOf(',')), 16);
     int count = Integer.parseInt(data.substring(data.indexOf(',') + 1), 16);
     try {
       byte value[] = new byte[count * 2];
